refactor: update blocks and util components (#251)
